EONEXT-129 - Changed the way to define development options for extended fields.
