+++ .env.example
@@ -5,3 +5,6 @@ FBS_BASEURL=https://fbs-openplatform.inlead.dev
 CMS_BASEURL=http://cms-mock.docker
 GRAPHQL_API_BASEURL=https://stg.lms.inlead.ws/naesbib/graphql
 COVERS_BASEURL=https://cover.dandigbib.org
+# Add the "data-eonext-ext-covers" and "data-eonext-ext-fields" to the document body
+# from the dpl-react/src/core/utils/config.tsx
+USE_DEVELOPMENT_OPTIONS=false
